+++ modules/interpreter/src/cpp/VertCatOperator.cpp
@@ -28,6 +28,17 @@
 //=============================================================================
 namespace Nelson {
 //=============================================================================
+static bool
+hasCell(const ArrayOfVector& v)
+{
+    for (indexType k = 0; k < v.size(); ++k) {
+        if (v[k].isCell()) {
+            return true;
+        }
+    }
+    return false;
+}
+//=============================================================================
 ArrayOf
 VertCatOperator(Evaluator* eval, const ArrayOfVector& v)
 {
@@ -40,11 +51,18 @@ VertCatOperator(Evaluator* eval, const ArrayOfVector& v)
         res = v[0];
     } break;
     default: {
+        bool asCell = hasCell(v);
         res = v[0];
         res.ensureSingleOwner();
         for (size_t k = 1; k < v.size(); k++) {
             ArrayOf arg2 = v[k];
-            res = eval->doBinaryOperatorOverload(res, arg2, VertCat, "vertcat");
+            if (asCell && (!arg2.isCell() && !arg2.isEmpty())) {
+                ArrayOf arg = ArrayOf::toCell(arg2);
+                res = eval->doBinaryOperatorOverload(
+                    res, arg, VertCat, "vertcat");
+            } else {
+                res = eval->doBinaryOperatorOverload(res, arg2, VertCat, "vertcat");
+            }
         }
     } break;
     }